add typed workflow client and input resources
